Extract parse method and NEW_LINE const
